Add db:connect command
