+++ examples/starter/template/db/connect.js
@@ -1,0 +1,3 @@
+const { DATABASE_URL } = require('./util.js')
+const { spawn } = require('child_process')
+spawn(`psql ${DATABASE_URL}`, [], { cwd: __dirname, stdio: 'inherit', shell: true })
